+++ packages/shared/src/utils/invoice/index.js
@@ -1,3 +1,4 @@
 export * from './invoice';
+export * from './invoiceItem';
 export * from './payments';
 export * from './display';
